Add default to hook implementation
